refactor: 파일 분리
- Credit과 Student객체를 정의하는 파일을 각각 분리
